pyproject.toml
[tool.mypy]
plugins = "sqlalchemy.ext.mypy.plugin"

[[tool.mypy.overrides]]
module = "furl.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "jwcrypto.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "uvicorn"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "uvicorn.logging.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "babel.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "postmarker.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sendgrid.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "dramatiq.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "passlib.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "asyncpg.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pymysql.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pycountry.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sentry_dramatiq.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "wtforms.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "apscheduler.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sqlalchemy_utils.*"
ignore_missing_imports = true

[tool.ruff]
extend-select = ["I"]
ignore = ["E501"]
per-file-ignores = {"conftest.py" = ["E731"], "fief/repositories/*" = ["E711", "E712"], "fief/alembic/*" = ["F401"]}

[tool.coverage.run]
concurrency = ["thread", "gevent"]
omit = ["fief/alembic/*"]

[tool.pytest.ini_options]
addopts = "--cov fief/ --cov-report=term-missing"
markers = ["test_data", "workspace_host", "access_token", "authenticated_admin", "htmx"]
asyncio_mode = "auto"

[tool.hatch]

[tool.hatch.build]
include = [
  "/fief/**/*.py",
  "/fief/py.typed",
  "/fief/alembic.ini",
  "/fief/locale",
  "/fief/static",
  "/fief/templates",
  "/fief/services/email_template/templates",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
source = "regex_commit"
commit_extra_args = ["-e"]
path = "fief/__init__.py"

[tool.hatch.envs.default]
dependencies = [
  "asgi-lifespan",
  "beautifulsoup4",
  "black",
  "coverage[toml]",
  "gevent",
  "httpx",
  "mypy",
  "pytest",
  "pytest-cov",
  "pytest-asyncio",
  "pytest-mock",
  "ruff",
  "tryceratops",
  "types-python-slugify",
  "types-pytz",
]

[tool.hatch.envs.default.scripts]
test = "pytest {args}"
test-cov-xml = "pytest --cov fief/ --cov-report=xml"
lint = [
  "black . ",
  "ruff --fix .",
  "mypy fief/",
  "tryceratops fief/",
]
lint-check = [
  "black --check .",
  "ruff .",
  "mypy fief/",
  "tryceratops fief/",
]

[build-system]
requires = ["hatchling", "hatch-regex-commit"]
build-backend = "hatchling.build"

[project]
name = "fief-server"
authors = [
  { name = "François Voron", email = "contact@fief.dev" }
]
description = "Users and authentication management SaaS"
readme = "README.md"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3 :: Only",
]
dynamic = ["version"]
requires-python = ">=3.10"
dependencies = [
    "aiomysql ==0.1.1",
    "aiosqlite ==0.17.0",
    "alembic ==1.8.1",
    "apscheduler ==3.9.1.post1",
    "argon2_cffi ==21.3.0",
    "asgi-babel==0.7.0",
    "asyncpg ==0.27.0",
    "dramatiq[redis] ==1.12.3",
    "fastapi ==0.87.0",
    "fastapi-users[sqlalchemy] ==10.2.1",
    "fief-client ==0.14.6",
    "furl ==2.1.3",
    "greenlet ==1.1.3",
    "gunicorn ==20.1.0",
    "httpx-oauth ==0.10.2",
    "itsdangerous ==2.1.2",
    "Jinja2 ==3.1.2",
    "jwcrypto ==1.4.2",
    "loguru ==0.6.0",
    "phonenumbers >=8.12.48,<8.14",
    "postmarker ==1.0",
    "sendgrid ==6.9.7",
    "psycopg2 ==2.9.5",
    "pycountry >=22.3.5,<22.4",
    "pydantic[dotenv] ==1.10.2",
    "python-slugify ==6.1.2",
    "pytz ==2022.2.1",
    "rich ==12.6.0",
    "sentry_dramatiq ==0.3.2",
    "sentry-sdk ==1.10.1",
    "sqlalchemy[mypy] ==1.4.44",
    "sqlalchemy-utils ==0.38.3",
    "typer ==0.7.0",
    "uvicorn[standard]==0.19.0",
    "WTForms==3.0.1",
]

[project.scripts]
fief = "fief.cli:app"

[project.urls]
Documentation = "https://docs.fief.dev"
Source = "https://github.com/fief-dev/fief"
